+++ experimental/canvaskit/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+ - Potential bug in `ready()` if already loaded.
 
 ## [0.3.1] - 2019-01-04
 ### Added